chore: make sure auth object is set immediately
