+++ packages/tanstack-react-start/src/server/middlewareHandler.ts
@@ -1,13 +1,16 @@
+import { AuthStatus, constants } from '@clerk/backend/internal';
+import { handleNetlifyCacheInDevInstance } from '@clerk/shared/netlifyCacheHandler';
 import type { AnyRouter } from '@tanstack/react-router';
 import {
   type CustomizeStartHandler,
   getEvent,
+  getWebRequest,
   type HandlerCallback,
   type RequestHandler,
 } from '@tanstack/react-start/server';
 
+import { errorThrower } from '../utils';
 import { authenticateRequest } from './authenticateRequest';
-import { ClerkHandshakeRedirect } from './errors';
 import { loadOptions } from './loadOptions';
 import type { LoaderOptions } from './types';
 import { getResponseClerkState } from './utils';
@@ -16,45 +19,51 @@ export function createClerkHandler<TRouter extends AnyRouter>(
   eventHandler: CustomizeStartHandler<TRouter>,
   clerkOptions: LoaderOptions = {},
 ) {
-  return (cb: HandlerCallback<TRouter>): RequestHandler => {
+  return async (cb: HandlerCallback<TRouter>): Promise<RequestHandler> => {
+    const request = getWebRequest();
+    const event = getEvent();
+    const loadedOptions = loadOptions(request, clerkOptions);
+
+    const requestState = await authenticateRequest(request, {
+      ...loadedOptions,
+      acceptsToken: 'any',
+    });
+
+    // Set auth object here so it is available immediately in server functions via getAuth()
+    event.context.auth = () => requestState.toAuth();
+
     return eventHandler(async ({ request, router, responseHeaders }) => {
-      const event = getEvent();
-
-      try {
-        const loadedOptions = loadOptions(request, clerkOptions);
-
-        const requestState = await authenticateRequest(request, {
-          ...loadedOptions,
-          acceptsToken: 'any',
+      const locationHeader = requestState.headers.get(constants.Headers.Location);
+      if (locationHeader) {
+        handleNetlifyCacheInDevInstance({
+          locationHeader,
+          requestStateHeaders: requestState.headers,
+          publishableKey: requestState.publishableKey,
         });
 
-        // Set context for getAuth to use
-        event.context.auth = () => requestState.toAuth();
-
-        const { clerkInitialState, headers } = getResponseClerkState(requestState, loadedOptions);
-
-        // Merging the TanStack router context with the Clerk context and loading the router
-        router.update({
-          context: { ...router.options.context, clerkInitialState },
+        return new Response(null, {
+          status: 307,
+          headers: requestState.headers,
         });
-
-        headers.forEach((value, key) => {
-          responseHeaders.set(key, value);
-        });
-
-        await router.load();
-      } catch (error) {
-        if (error instanceof ClerkHandshakeRedirect) {
-          // returning the response
-          return new Response(null, {
-            status: error.status,
-            headers: error.headers,
-          });
-        }
-
-        // rethrowing the error if it is not a Response
-        throw error;
       }
+
+      if (requestState.status === AuthStatus.Handshake) {
+        // eslint-disable-next-line @typescript-eslint/only-throw-error
+        throw errorThrower.throw('Clerk: unexpected handshake without redirect');
+      }
+
+      const { clerkInitialState, headers } = getResponseClerkState(requestState, loadedOptions);
+
+      // Merging the TanStack router context with the Clerk context and loading the router
+      router.update({
+        context: { ...router.options.context, clerkInitialState },
+      });
+
+      headers.forEach((value, key) => {
+        responseHeaders.set(key, value);
+      });
+
+      await router.load();
 
       return cb({ request, router, responseHeaders });
     });
